requirements.txt
Django==1.10.6
django-registration==2.2